Fixes, changes and hygiene
Changes:
* NodePath now knows 2 new ways of referencing node/resources -
** referencing global resource: `.:@resource/path.ext:` which allows animating material properties for all users of that material/etc.
** auto-locating paths `|node_name/subnode:sub_path` when resolved will recursively search for the given `node_name`

Fixes:
* Signal connections fixed in a few editor plugins
* A bug in input processing when a modal window poped-up was fixed.
* A few printf-formats were fixed
* fix node-name filter in scene tree dock.
* fix undo-redo operation type for eastl::function
* re-add _request_gizmo to exposed node_3d editor methods.
* Curve2D::_bake was broken, no longer is :)

Hygiene:
* A bunch of pedantic-level warnings were fixed + some single-line if/else statements were surrounded with '{}'

+++ core/io/logger.cpp
@@ -69,11 +69,11 @@ void Logger::log_error(StringView p_function, StringView p_file, int p_line, Str
     else
         err_details = p_code;
 
-    logf_error(FormatVE("%s: %.*s\n",err_type,err_details.size(),err_details.data()));
+    logf_error(FormatVE("%s: %.*s\n",err_type,(int)err_details.size(),err_details.data()));
     logf_error(FormatVE("   at: %.*s (%.*s:%i) - %.*s\n",
-                        p_function.size(),p_function.data(),
-                        p_file.size(),p_file.data(),p_line,
-                        p_code.size(),p_code.data())
+                        (int)p_function.size(),p_function.data(),
+                        (int)p_file.size(),p_file.data(),p_line,
+                        (int)p_code.size(),p_code.data())
                );
 }
